CORE-7115 Move ConsensualSignedTransaction from ledger-consensual-data to ledger-consensual-flow
