+++ components/ledger/ledger-consensual-flow/src/main/java/net/corda/ledger/consensual/flow/impl/transaction/package-info.java
@@ -1,0 +1,4 @@
+@Export
+package net.corda.ledger.consensual.flow.impl.transaction;
+
+import org.osgi.annotation.bundle.Export;
