+++ components/ledger/ledger-consensual-flow/src/test/kotlin/net/corda/ledger/consensual/flow/impl/transaction/serializer/amqp/ConsensualSignedTransactionSerializerTest.kt
@@ -1,4 +1,4 @@
-package net.corda.ledger.consensual.data.transaction.serializer.amqp
+package net.corda.ledger.consensual.flow.impl.transaction.serializer.amqp
 
 import net.corda.application.impl.services.json.JsonMarshallingServiceImpl
 import net.corda.cipher.suite.impl.CipherSchemeMetadataImpl
@@ -7,6 +7,7 @@ import net.corda.crypto.merkle.impl.MerkleTreeProviderImpl
 import net.corda.internal.serialization.amqp.helper.TestSerializationService
 import net.corda.ledger.common.data.transaction.serializer.amqp.WireTransactionSerializer
 import net.corda.ledger.common.testkit.mockSigningService
+import net.corda.ledger.consensual.flow.impl.transaction.amqp.ConsensualSignedTransactionSerializer
 import net.corda.ledger.consensual.testkit.getConsensualSignedTransactionExample
 import net.corda.v5.application.serialization.deserialize
 import org.junit.jupiter.api.Assertions
@@ -19,14 +20,20 @@ class ConsensualSignedTransactionSerializerTest {
     private val digestService = DigestServiceImpl(cipherSchemeMetadata, null)
     private val merkleTreeProvider = MerkleTreeProviderImpl(digestService)
     private val jsonMarshallingService = JsonMarshallingServiceImpl()
-    private val serializationServiceNullCfg = TestSerializationService.getTestSerializationService({}, cipherSchemeMetadata)
+    private val serializationServiceNullCfg =
+        TestSerializationService.getTestSerializationService({}, cipherSchemeMetadata)
     private val serializationService = TestSerializationService.getTestSerializationService({
-        it.register(WireTransactionSerializer(
-            merkleTreeProvider,
-            digestService,
-            jsonMarshallingService
-        ), it)
-        it.register(ConsensualSignedTransactionSerializer(serializationServiceNullCfg, mockSigningService(), mock()), it)
+        it.register(
+            WireTransactionSerializer(
+                merkleTreeProvider,
+                digestService,
+                jsonMarshallingService
+            ), it
+        )
+        it.register(
+            ConsensualSignedTransactionSerializer(serializationServiceNullCfg, mockSigningService(), mock()),
+            it
+        )
     }, cipherSchemeMetadata)
 
     @Test
