+++ libs/ledger/ledger-consensual-data/build.gradle
@@ -1,7 +1,6 @@
 plugins {
     id 'corda.common-publishing'
     id 'corda.common-library'
-    id 'corda.osgi-test-conventions'
 }
 
 description 'Corda ledger consensual data'
@@ -14,28 +13,6 @@ dependencies {
     implementation "net.corda.kotlin:kotlin-stdlib-jdk8-osgi"
     implementation "net.corda:corda-ledger-consensual"
 
-    implementation project(":libs:serialization:serialization-internal")
     implementation project(":libs:crypto:crypto-core")
     implementation project(":libs:ledger:ledger-common-data")
-
-    testImplementation "org.jetbrains.kotlin:kotlin-test:$kotlinVersion"
-    testImplementation "org.mockito.kotlin:mockito-kotlin:$mockitoKotlinVersion"
-
-    testImplementation "net.corda:corda-ledger-common"
-    testImplementation project(':testing:test-serialization')
-    testImplementation project(':testing:ledger:ledger-consensual-testkit')
-    testImplementation project(":libs:serialization:serialization-amqp")
-    testImplementation project(":libs:crypto:cipher-suite-impl")
-    testImplementation project(":libs:crypto:merkle-impl")
-    testImplementation project(":libs:application:application-impl")
-
-    integrationTestImplementation project(':testing:sandboxes')
-
-    integrationTestRuntimeOnly project(':libs:crypto:cipher-suite-impl')
-    integrationTestRuntimeOnly project(":libs:lifecycle:lifecycle-impl")
-    integrationTestRuntimeOnly project(":libs:flows:session-manager-impl")
-    integrationTestRuntimeOnly project(':libs:messaging:db-message-bus-impl')
-    integrationTestRuntimeOnly project(':libs:messaging:messaging-impl')
-    integrationTestRuntimeOnly project(':libs:serialization:serialization-kryo')
-    integrationTestRuntimeOnly "org.apache.aries.spifly:org.apache.aries.spifly.dynamic.framework.extension:$ariesDynamicFrameworkExtensionVersion"
 }
