--- libs/ledger/ledger-consensual-data/test.bndrun
@@ -1,13 +1,0 @@
--tester: biz.aQute.tester.junit-platform
--runee: JavaSE-11
--runfw: org.apache.felix.framework
--resolve.effective: resolve,active
--runtrace: true
-
--runrequires: \
-    bnd.identity;id='net.corda.ledger-consensual-data',\
-    bnd.identity;id='${project.archivesBaseName}-tests',\
-    bnd.identity;id='net.corda.ledger-common-data',\
-    bnd.identity;id='junit-jupiter-engine',\
-    bnd.identity;id='junit-platform-launcher',\
-    bnd.identity;id='slf4j.simple'
